requirements.txt
ahk==1.7.6
ahk-binary==2023.9.0
backoff==2.2.1
openai==1.2.3
python-dotenv==1.0.0
numpy==1.24.3
tiktoken==0.5.1
colorama
pyautogui==0.9.54
pydirectinput==1.0.4; sys_platform == "win32"
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
pillow>=10.3.0
Multi-Template-Matching==1.6.6
imageio[ffmpeg]
mss==9.0.1
aiohttp
easyocr==1.7.1
spacy==3.7.2
chardet==5.2.0
matplotlib==3.8.4
supervision==0.21.0
pyobjc-framework-Quartz==10.0; sys_platform == "darwin"
pyobjc-framework-Cocoa==10.0; sys_platform == "darwin"
git+https://github.com/facebookresearch/segment-anything.git
pywin32; sys_platform == "win32"
dataclass_wizard==0.22.3
dill==0.3.8
anthropic
